型ユーティリティ追加
