+++ src/theme/buttonTheme/variants.ts
@@ -1,0 +1,22 @@
+export const buttonChildrenVariants = {
+  isLoading: {
+    true: "",
+    false: "",
+  },
+  background: {
+    filled: "",
+    white: "",
+  },
+  theme: {
+    white: "text-primary-600",
+    primary: "text-white",
+    "no-background": "text-primary-600 active:text-primary-900",
+    red: "",
+    gray: "",
+  },
+  size: {
+    small: "text-xs",
+    medium: "text-sm",
+    large: "text-base",
+  },
+};
